Unittest framework ready
* running runtest.py now runs all unittest.TestCase specified in
modules test*.py
* Fixed some bugs in cancel
* Added sbatch options to interface

+++ sutils/slurm_interface/api.py
@@ -5,26 +5,36 @@ import getpass
 import subprocess
 
 from . import config
+from ..test import test
+from ..utils.util import to_list
 
 def run_command(cmd, args):
-    p = subprocess.Popen(args, executable=cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-    p.wait()
-    stdout, stderr = p.communicate()
-    p.terminate()
+    if test.testmode():
+        test.cmd_buffer.write(cmd + " " + " ".join(args))
+        stdout = ""
+        stderr = ""
+    else:
+        p = subprocess.Popen(args, executable=cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+        p.wait()
+        stdout, stderr = p.communicate()
+        p.terminate()
 
     retval = 0 if len(stderr) > 0 else 1
 
     return retval, stdout, stderr
 
-def sbatch(work_dir, *args):
+def sbatch(work_dir, *args, **kwargs):
     named_args = config.SbatchConfig(work_dir=work_dir).to_list()
     named_args += args # this is to support extra arguments. No guarantee!
+    for arg in kwargs:
+        named_args.append("--{}={}".format(arg, kwargs[arg]))
 
     retval, stdout, stderr = run_command('sbatch', named_args)
     return retval, stdout
 
 def scancel(job_id):
-    args = config.Scancel_Options(job_id=job_id).to_list()
+    #args = config.Scancel_Options(job_id=job_id).to_list()
+    args = to_list(job_id)
     retval, stdout, stderr = run_command('scancel', args)
     return 0
 
@@ -54,8 +64,12 @@ class Result(object):
 class SqueueResult(Result):
     def __init__(self, data):
         # convert to numpy array and use slices?
-        self.jobid = [int(d[0]) for d in data]
-        self.status = [d[1] for d in data]
+        if len(data[0]) > 0:
+            self.jobid = [int(d[0]) for d in data]
+            self.status = [d[1] for d in data]
+        else:
+            self.jobid = None
+            self.status = None
 
     def get(self, jobid):
         try:
